Fixed some signed/unsigned compares in sdf, usd and usdGeom.
Also disabled a "maybe uninitialized" warning in sdf.

(Internal change: 1973935)

+++ pxr/usd/lib/sdf/listEditor.h
@@ -234,8 +234,8 @@ protected:
         // We assume that duplicate data items are never allowed to be
         // authored. For full generality, this information ought to come from
         // the layer schema.
-        for (int i = 0; i < newValues.size(); ++i) {
-            for (int j = i + 1; j < newValues.size(); ++j) {
+        for (size_t i = 0; i < newValues.size(); ++i) {
+            for (size_t j = i + 1; j < newValues.size(); ++j) {
                 if (newValues[i] == newValues[j]) {
                     TF_CODING_ERROR("Duplicate item '%s' not allowed for "
                                     "field '%s' on <%s>",
